multimodal_emotion_detection/requirements.txt
deepface==0.0.79
opencv-python>=4.8.0
numpy>=1.24.0
librosa>=0.10.1
soundfile>=0.12.1
transformers>=4.36.0
torch>=2.1.0
scipy>=1.11.0
fastapi>=0.110.0
uvicorn>=0.23.0
pydantic>=2.4.0
# Added for training scripts
pandas>=2.1.0
scikit-learn>=1.3.0
# Added for image training
requests>=2.31.0
torchvision>=0.16.0
Pillow>=10.0.0